requirements.txt
torch==1.7.0
pandas==1.2.4
numpy==1.19.4
streamlit==0.80.0
python-mnist
